# packages/contracts/src/openapi.yaml
openapi: 3.1.0
info:
  title: Xenitex Appliance API
  version: 1.0.0
  description: >
    v1 of the on-premise appliance API. This document is the single source of
    truth (P1-01) — the TypeScript client and types in this package are
    generated from it, never hand-written (P1-02).

    Conventions (see docs/adr/0006-api-conventions.md, accepted at GATE 1):
    opaque cursor pagination on every collection (`?cursor=&limit=`, response
    `{ items, nextCursor }`); RFC 9457 Problem Details on every non-2xx
    response; `Idempotency-Key` required on every creation POST;
    `ETag`/`If-Match` required on every mutation of a mutable resource; URI
    path versioning (`/v1/...`); long-running operations return `202` with a
    `Location` header pointing at a pollable resource, never a blocking
    request.

    JSON casing: camelCase, matching packages/domain. Any field whose value
    originates from a scanned target (SEC-17/AIP-02 — banners, headers,
    certificate subjects, hostnames, identity-key values, OS/service/product
    inference) carries an `Untrusted` suffix in its property name. This is
    the same greppable-naming convention as the `_untrusted` column suffix in
    docs/adr/0007-schema-conventions.md, extended to the API boundary so a
    frontend consumer can never mistake such a value for first-party data —
    it must be escaped on output, never rendered as HTML, never interpreted.
servers:
  - url: /v1
...
security:
  - sessionCookie: []
paths:
...
  /assets:
    get:
      tags: [Assets]
      operationId: listAssets
      summary: List assets (virtualised list, faceted filters, P1-09/P1-22)
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: lifecycleState
          in: query
          schema: { $ref: '#/components/schemas/AssetLifecycleState' }
        - name: criticality
          in: query
          schema: { $ref: '#/components/schemas/AssetCriticality' }
        - name: exposure
          in: query
          schema: { $ref: '#/components/schemas/ExposureClassification' }
        - name: tag
          in: query
          schema: { type: string }
        - name: isFragile
          in: query
          schema: { type: boolean }
        - name: q
          in: query
          description: Free-text search over address/hostname/tags.
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/AssetListPage' }
  /assets/{assetId}:
    parameters: [{ $ref: '#/components/parameters/AssetIdPath' }]
    get:
      tags: [Assets]
      operationId: getAsset
      summary: Asset detail — identity keys, address/hostname history, services (P1-09)
      responses:
        '200':
          description: 'OK.'
          content: { application/json: { schema: { $ref: '#/components/schemas/Asset' } } }
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Assets]
      operationId: updateAsset
      summary: Update owner, criticality, tags
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ownerTeam: { oneOf: [{ type: string }, { type: 'null' }] }
                businessCriticality: { $ref: '#/components/schemas/AssetCriticality' }
                tags: { type: array, items: { type: string } }
      responses:
        '200':
          description: 'OK.'
          content: { application/json: { schema: { $ref: '#/components/schemas/Asset' } } }
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StaleResource' }
  /assets/{assetId}/issues:
    parameters: [{ $ref: '#/components/parameters/AssetIdPath' }]
    get:
      tags: [Assets]
      operationId: listAssetIssues
      summary: Open issues for this asset
      parameters:
        [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        '200': { $ref: '#/components/responses/IssueListPage' }
  /assets/{assetId}/scan-history:
    parameters: [{ $ref: '#/components/parameters/AssetIdPath' }]
    get:
      tags: [Assets]
      operationId: listAssetScanHistory
      summary: Scan runs that targeted this asset
      parameters:
        [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        '200': { $ref: '#/components/responses/ScanRunListPage' }
  /asset-merge-candidates:
    get:
      tags: [Assets]
      operationId: listAssetMergeCandidates
      summary: Low-confidence merge candidates awaiting operator review (MOD-06)
      parameters:
        [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        '200':
          description: 'OK.'
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/AssetMergeEvent' } }
                  nextCursor: { oneOf: [{ type: string }, { type: 'null' }] }
  /asset-merges:
    post:
      tags: [Assets]
      operationId: createAssetMerge
      summary: Operator-confirmed merge of two assets (MOD-06)
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [survivorAssetId, mergedAssetId]
              properties:
                survivorAssetId: { type: string }
                mergedAssetId: { type: string }
      responses:
        '201':
          {
            description: 'Created.',
            content:
              { application/json: { schema: { $ref: '#/components/schemas/AssetMergeEvent' } } },
          }
        '400': { $ref: '#/components/responses/ValidationError' }
  /asset-merge-events/{eventId}/reverse:
    parameters:
      - name: eventId
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Assets]
      operationId: reverseAssetMerge
      summary: Reverse an incorrect merge (MOD-06) — restores the merged asset to active, keeps its history
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        '200':
          {
            description: 'OK.',
            content:
              { application/json: { schema: { $ref: '#/components/schemas/AssetMergeEvent' } } },
          }
        '404': { $ref: '#/components/responses/NotFound' }

  # --------------------------------------------------------- AssetGroups
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: xenitex_session
      description: >
        Opaque server-side session (SEC-08). State-changing requests also
        require the double-submit CSRF token in the `X-CSRF-Token` header,
        matched against the session's csrfTokenHash server-side.
  parameters:
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from a previous page's `nextCursor`. Never a raw offset (ADR 0006).
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Current ETag of the resource. A stale value returns 409 (ADR 0006).
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated key. A retried request with the same key returns the original response rather than creating a second resource (ADR 0006).
      schema: { type: string }
...
    AssetIdPath:
      name: assetId
      in: path
      required: true
      schema: { type: string }
...
  headers:
    ETag:
      description: Current entity tag; send back in `If-Match` on the next mutation.
      schema: { type: string }
...
  responses:
...
    NotFound:
      description: No such resource.
      headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
      content:
        { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
...
    StaleResource:
      description: 'code: concurrency.stale_resource — the supplied If-Match no longer matches (ADR 0006).'
      headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
      content:
        { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
...
    AssetListPage:
      description: Page of assets.
      content:
        application/json:
          schema:
            type: object
            required: [items, nextCursor]
            properties:
              items: { type: array, items: { $ref: '#/components/schemas/Asset' } }
              nextCursor: { oneOf: [{ type: string }, { type: 'null' }] }
    IssueListPage:
      description: Page of issues.
      content:
        application/json:
          schema:
            type: object
            required: [items, nextCursor]
            properties:
              items: { type: array, items: { $ref: '#/components/schemas/Issue' } }
              nextCursor: { oneOf: [{ type: string }, { type: 'null' }] }
    ScanRunListPage:
      description: Page of scan runs.
      content:
        application/json:
          schema:
            type: object
            required: [items, nextCursor]
            properties:
              items: { type: array, items: { $ref: '#/components/schemas/ScanRun' } }
              nextCursor: { oneOf: [{ type: string }, { type: 'null' }] }
  schemas:
...
    AssetLifecycleState: { type: string, enum: [active, inactive, decommissioned, merged] }
    AssetCriticality: { type: string, enum: [low, medium, high, critical] }
    ExposureClassification: { type: string, enum: [internal, dmz, external, isolated, unknown] }
...
    Asset:
      type: object
      required:
        - id
        - lifecycleState
        - identityKeys
        - addresses
        - hostnames
        - services
        - businessCriticality
        - exposureClassification
        - tags
        - isFragile
        - firstSeen
        - lastSeen
        - riskRating
        - riskBand
      properties:
        id: { type: string }
        lifecycleState: { $ref: '#/components/schemas/AssetLifecycleState' }
        mergedIntoAssetId: { oneOf: [{ type: string }, { type: 'null' }] }
        identityKeys: { type: array, items: { $ref: '#/components/schemas/AssetIdentityKey' } }
        addresses: { type: array, items: { $ref: '#/components/schemas/AssetAddressHistoryEntry' } }
        hostnames:
          { type: array, items: { $ref: '#/components/schemas/AssetHostnameHistoryEntry' } }
        osInference:
          oneOf:
            - $ref: '#/components/schemas/OsInference'
            - type: 'null'
        services: { type: array, items: { $ref: '#/components/schemas/AssetService' } }
        ownerTeam: { oneOf: [{ type: string }, { type: 'null' }] }
        businessCriticality: { $ref: '#/components/schemas/AssetCriticality' }
        exposureClassification: { $ref: '#/components/schemas/ExposureClassification' }
        tags: { type: array, items: { type: string } }
        isFragile: { type: boolean }
        firstSeen: { type: string, format: date-time }
        lastSeen: { type: string, format: date-time }
        riskRating:
          {
            oneOf: [{ type: number, minimum: 0, maximum: 100 }, { type: 'null' }],
            description: 'docs/issues-scoring-dashboard-spec.md SCORE 2.4/SCORE-08 — null until the asset has at least one open issue.',
          }
        riskBand:
          oneOf:
            - $ref: '#/components/schemas/RiskBand'
            - type: 'null'
    AssetMergeEvent:
      type: object
      required: [id, survivorAssetId, mergedAssetId, policyVersion, confidence, performedAt]
      properties:
        id: { type: string }
        survivorAssetId: { type: string }
        mergedAssetId: { type: string }
        matchedIdentityKeyType: { $ref: '#/components/schemas/IdentityKeyType' }
        policyVersion: { type: integer }
        confidence: { type: number, minimum: 0, maximum: 1 }
        performedBy: { oneOf: [{ type: string }, { type: 'null' }] }
        performedAt: { type: string, format: date-time }
        reversedBy: { oneOf: [{ type: string }, { type: 'null' }] }
        reversedAt: { oneOf: [{ type: string, format: date-time }, { type: 'null' }] }
        reversalReason: { oneOf: [{ type: string }, { type: 'null' }] }
...